Add the ILogger interface and FileLogger implementation
This change introduces the ILogger interface to abstract writing log
messages to an arbitrary sink so that this interface can be passed to
other components for logging behavior.  It also adds the FileLogger
implementation which is a refactored LogWriter implementation.
